+++ extension.js
@@ -1,7 +1,7 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 const vscode = require('vscode');
-const { registerGetALRecordFields } = require('./src/AIToolsDefinition.js');
+const { registerGetALRecordFields, registerGetALObjectDefinition, registerOpenALObjectInEditor } = require('./src/AIToolsDefinition.js');
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 
@@ -10,6 +10,8 @@ const { registerGetALRecordFields } = require('./src/AIToolsDefinition.js');
  */
 function activate(context) {
 	registerGetALRecordFields(context);
+	registerGetALObjectDefinition(context);
+	registerOpenALObjectInEditor(context);
 	let disposableSelection = vscode.commands.registerCommand('JALVarNaming.AlVarNameSel', function () {
 		// The code you place here will be executed every time your command is executed
 		// Display a message box to the user
@@ -37,6 +39,12 @@ function activate(context) {
 		});
 	});
 	context.subscriptions.push(disposableAll);
+
+	let disposablePrueba = vscode.commands.registerCommand('JALVarNaming.Prueba', function () {
+			//call get AlDefinition
+			logSalesHeader();
+	});
+	context.subscriptions.push(disposablePrueba);
 
 	let disposableCatchDocumentChanges = vscode.commands.registerCommand('JALVarNaming.CatchDocumentChanges', function () {
 		const rename = require('./src/RenameVars.js');
@@ -140,6 +148,14 @@ function activate(context) {
 		extractEvents.setNewEventSubsFile();
 	});
 	context.subscriptions.push(disposablesetNewEventSubsFile);
+
+
+	async function logSalesHeader() {
+			const AIToolsDefinition = require('./src/AIToolsDefinition.js');
+			console.log(await AIToolsDefinition.GetALDefinition('Sales Header','table',36));
+
+}
+
 }
 // @ts-ignore
 exports.activate = activate;